creative_flow/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
requests>=2.31.0
beautifulsoup4>=4.12.0
